Split TableCellRenderes from TableCellEditors
This rationalization helps to better follow the swing abstractions
of table models and increase separation of concerns.

(WIP: ContributedPlatforms needs a similar refactoring that will be
done in the next commits)
